Make Bash print the commands, echo all variables, fix URL and fix DATA substitution
It seems that I am getting some super weird values for the NAME, so I am printing everything out step by step to see what's up.

The API URL was missing a slash which I have now added.

The DATA variable was using single quotes making the string a literal. I fixed it so that string concatenation is used and I get the DATE variable properly interpolated to the string.