Make the workflow actually make the cURL call
Everything should be set up now, so let's let this go and see what happens. I am still providing my handle in the cURL -u option value, so maybe the commit will be associated to me. I believe there should be a way to provide just the PAT though, because in REST calls I make elsewhere in code, I was able to pass "Authorization: `token ${token}`" to the request headers and GitHub was fine with that. I might look into doing the same here.