Rename step, use github context field for handle, drop response header printout
I have not renamed the step since it was just printing a test echo, so here it is updated to reflect the changes.

I have changed the hardcoded GitHub handle to the GitHub context object field value. This is not consequential for me in any way, but for anyone copy-pasting it will make the workflow more solid.

I have dropped the response header printout as it doesn't contain anything too interesting over the response body. It does include the integration PAT rate limits, but those I already know are a 1000 and the current used up amount is always low in the case of my repos, so I don't worry too much about this.