Quote the keys of branch and fix structure of message
I am not sure if the keys need to be quoted or not. The message I got now says "InputObject 'CommittableBranch' doesn't accept argument 'nameWithOwner'". Does that mean the keys are needed? I don't know. I quoted them again to see if that changes things. And the message is also a complex object, not a string, so I fixed that, too.