Capture the GraphQL response and learnings about the mystery number
With this I have verified GraphQL works for multiple files so even if my experiments in using the Git client later on don't work out, I am golden, because all my use-cases can now be covered by this. GraphQL can be used in any case, REST in a simple case of a single file where GraphQL would be more chatty.