Clear out the Git CLI based tasks that I've completed
It is possible to configure the Git identity such that it appears to be coming from the GitHub Actions service account by using its initials. This needs to be done explicitly as opposed to the API based methods where it is derived from the PAT though, so it is less resilient to changes.